hangman edit

+++ src/SanviSaraCalculator/ArgData.java
@@ -1,31 +1,33 @@
+//Console Based Calculator Created by Sanvi
+//Model
 package SanviSaraCalculator;
-
+//AP req- class for control
 public class ArgData {
-    private int arg1;
-    private int arg2;
-    private int op;
+    private double arg1;
+    private double arg2;
+    private double op;
 
-    public int getArg1() {
+    public double getArg1() {
         return arg1;
     }
 
-    public int getArg2(){
+    public double getArg2(){
         return arg2;
     }
 
-    public int getOp(){
+    public double getOp(){
         return op;
     }
 
-    public void setArg1(int arg1) {
+    public void setArg1(double arg1) {
         this.arg1 = arg1;
     }
 
-    public void setArg2(int arg2){
+    public void setArg2(double arg2){
         this.arg2 = arg2;
     }
 
-    public void setOp(int op){
+    public void setOp(double op){
         this.op = op;
     }
 }
